pyproject.toml
[tool.poetry]
name = "ray-app"
version = "0.1.0"
description = ""
authors = ["Ainaz Kabirov <grimoruu@yandex.ru>"]
readme = "README.md"
packages = [{ include = "ray_app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
alembic = "^1.8.1"
psycopg2-binary = "^2.9.5"
ruff = "^0.0.247"
ndicts = "^0.3.0"
pyjwt = "^2.6.0"
pydantic = {extras = ["email"], version = "^1.10.2"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
black = "^23.1.0"
mypy = "^1.0.0"
pyyaml = "^6.0"
asyncpg = "^0.27.0"
redis = "^4.5.5"
orjson = "^3.9.1"
aio-pika = "^9.2.1"
aiosmtplib = "^2.0.2"
phonenumbers = "^8.13.19"
itsdangerous = "^2.1.2"
sqlalchemy = {version = "1.4.46", extras = ["mypy"]}
requests = "^2.31.0"
aiohttp = "^3.9.1"
arq = "^0.25.0"
httpx = "^0.25.2"
setuptools = "^69.0.2"
fastapi-utils = "^0.2.1"
types-redis = "^4.6.0.11"
apscheduler = "^3.10.4"
schedule = "^1.2.1"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
types-requests = "^2.31.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
select = ["F", "E", "I", "U", "N", "C", "B", "A", "T", "Q", "ANN", "RUF"]
ignore = ["A003", "ANN101", "ANN401", "Q000", "E902", "B008", "B904", "N805", "RUF001"]

exclude = [".venv", ".db", "migrations/versions"]

[tool.mypy]
plugins = "sqlalchemy.ext.mypy.plugin"
ignore_missing_imports = true

exclude = [".venv", ".db", "migrations/versions", "test.py", "tests/pytest"]

[[tool.mypy.overrides]]
module = ["ndicts.*"]
ignore_missing_imports = true

[tool.black]
line-length = 120

exclude = '''
^/(
  (
    | \.git          # root of the project
    | \.venv
    | \.db
    | \.migrations\versions

  )/
)
'''

[tool.mypy-yaml]
ignore_missing_imports = true